bootstrap tipp-addon
first changes on the addon tippspiel for bootstrap.
some files are a pain in the ass and there may be no choice but to keep the table-structure beacause of to many informations to show

+++ lmo/addon/tipp/lmo-tippcalceinsicht.php
@@ -114,7 +114,7 @@ if (!file_exists($einsichtfile)) {
     $dum = explode('|', $tippdaten[$i-1]);
     $op1 = $dum[0];
     // Nick
-    $op3 = substr($dum[1], 2)-1;
+    $op3 = (int)substr($dum[1], 2)-1;
     // Spieltagsnummer
     $op4 = substr($dum[1], 0, 2);
     // TP
